requirements.txt
streamlit
psycopg2
python-dotenv
redis
tensorflow
keras